pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "inference-latency-study"
version = "0.1.0"
description = "Controlled benchmark for inference latency under agent-like payloads."
requires-python = ">=3.10"
dependencies = [
  "openai>=1.0.0",
  "PyYAML>=6.0",
  "pandas>=2.0.0",
  "Pillow>=10.0.0",
  "numpy>=1.24.0",
  "torch>=2.3.0",
  "torchvision>=0.18.0",
  "transformers==4.52.4",
]

[project.scripts]
ils = "bench.cli:main"

[tool.setuptools]
packages = ["bench", "bench.backends", "bench.infra", "bench.payloads", "analysis"]
